se restringieron el numero de decimales que despliega la app en cuanto el angulo

+++ appBike/main.py
@@ -252,7 +252,7 @@ class App(MDApp):
                 acceleration = float(acceleration)
                 print(f"acceleration-> {acceleration}")
                 await self.dataTx_queue.put(json.dumps({'acceleration_y': acceleration}))
-                anguloy = math.asin(acceleration / 9.81) * (180.0 / math.pi)
+                anguloy = int(math.asin(acceleration / 9.81) * (180.0 / math.pi))
                 self.an_button.text = f'Angle : {anguloy}°'
             except Exception as e:
                 print(f'Exception in angle:: {e}')
